Change function to use 'long' type

+++ src/main/java/io/bluebeaker/questtweaker/ctintegration/FunctionManager.java
@@ -25,7 +25,7 @@ public class FunctionManager {
         return FUNCTIONS.get(functionID);
     }
 
-    public static int runFunction(String functionID,EntityPlayer player){
+    public static long runFunction(String functionID,EntityPlayer player){
         IQuestsFunction function = getFunction(functionID);
         if(function==null)
             return -1;
